separate RHI codes
