+++ got2d/CMakeLists.txt
@@ -56,6 +56,8 @@ if(MSVC)
 list(APPEND GOT2D_RHI_INCLUDE_FILES 
 	RHI/dx11/inner_RHI.h
 	RHI/dx11/RHI.cpp
+	RHI/dx11/rhi_device.cpp
+	RHI/dx11/rhi_context.cpp
 )
 source_group(RHI_DX11 FILES ${GOT2D_RHI_INCLUDE_FILES})
 else()
